backend/requirements.txt
# HAL Backend Dependencies

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client (for Ollama)
httpx==0.26.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1

# Utilities
python-dotenv==1.0.0

# Async file operations
aiofiles==23.2.1

# Resource monitoring
psutil==5.9.8
